Move user_follow to separate file
Also add unit tests
