+++ models/fixtures/follow.yml
@@ -1,0 +1,4 @@
+-
+  id: 1
+  user_id: 4
+  follow_id: 2
